requirements.txt
matplotlib==3.5.3
numpy==1.24.2
pandas==1.3.5
Pillow==9.4.0
scikit_learn==1.2.1
seaborn==0.11.2
tensorflow==2.11.0
grpcio==1.32.0